backend/requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==3.2.19
djangorestframework==3.14.0
importlib-metadata==4.2.0
isort==5.11.5
mccabe==0.7.0
Pillow==9.5.0
pycodestyle==2.10.0
pyflakes==2.5.0
python-barcode==0.14.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.6.3
zipp==3.15.0
